Removed contents from home.md, added News. Refs #14.
The homepage will show a description of Copilot and a short example in code. In
the future these two should be provided by their own .md files, that will fill
parts of a template. For now these parts will be hardcoded in home.tpl. The
rest of the page will be used a news section.